Refactor common crate * Add SlurpError * Refactor slurp_req, slurp_url so they return SlurpResult * Replace slurp functions into native/wasm common::transport * Replace native executor mod into native_executor.rs * Rename executor.rs into wasm_executor.rs * Replace wio mod from common.rs into wio.rs
